+++ src/api/wgl/mod.rs
@@ -79,7 +79,8 @@ impl Context {
     /// # Unsafety
     ///
     /// The `window` must continue to exist as long as the resulting `Context` exists.
-    pub unsafe fn new(pf_reqs: &PixelFormatRequirements, opengl: &GlAttributes<HGLRC>,
+    #[inline]
+    pub fn new(pf_reqs: &PixelFormatRequirements, opengl: &GlAttributes<HGLRC>,
                       window: HWND) -> Result<Context, CreationError>
     {
         let hdc = GetDC(window);