+++ src/platform/emscripten/mod.rs
@@ -75,7 +75,7 @@ impl Context {
 
     /// See the docs in the crate root file.
     #[inline]
-    pub unsafe fn new_separate(
+    pub fn new_separate(
         _window: &winit::Window,
         _events_loop: &winit::EventsLoop,
         _pf_reqs: &PixelFormatRequirements,