+++ src/platform/macos/mod.rs
@@ -42,6 +42,7 @@ pub struct HeadlessContext {
 }
 
 impl Context {
+    #[inline]
     pub fn new(
         window_builder: winit::WindowBuilder,
         events_loop: &winit::EventsLoop,
@@ -161,7 +162,7 @@ impl Context {
 
     /// See the docs in the crate root file.
     #[inline]
-    pub unsafe fn new_separate(
+    pub fn new_separate(
         _window: &winit::Window,
         _events_loop: &winit::EventsLoop,
         _pf_reqs: &PixelFormatRequirements,